TokenIndex for Section -> token retrieval (#21)
This PR fixes #17 

## Implementation details

This collection gathers all the tokens that compose a Section. The functionality will help linting by block instead of dealing with the entirety of tokens.
The function `getTokensBySectionId` will be empty if AST property hasn't been called. It is implemented this way, given that the IDs are only present when the tokens -> AST are parsed. 

In order to use it. you can take a look at the test `Simpler section to tokens.` which exposes a simple way to use it. 

in the nutshell:

```typescript

 import { FluentBitSchema } from '@calyptia/fluent-bit-config-parser';

  // The file should exist in your filesystem if is relative. In this case given that is fullPath it can be fictitious. 
 const filePath = '/nowhere/ephemeral.conf';
    const rawConfig = `
    [INPUT]
      Name        tail # some comment
      Tag         tail.01
      Path        /var/log/system.log
    `;

    const config = new FluentBitSchema(rawConfig, filePath);
    const ast = config.AST;
    // sections appear in order, so the INPUT section is in the 0 index. 
    const [inputSection] = ast;

   const tokens =  config.getTokensFromSectionId(inputSection.id);

  // do something fun with tokens. 
```

I've also reorganized the sections in tests, just to keep things more cohesive. 
![image](https://user-images.githubusercontent.com/226559/153078322-e6d73deb-2c68-4a52-a696-78b90b211160.png)