Absolute paths options + handling absolute paths in includes. (#32)
Many configurations might have full paths on their includes.  This could be the case given how confusing relative paths become. When these errors happen inside fluent bit the error message is a bit obscure and so it might lead to the overuse of full path for more clarification. 

The parser provides better guidance and line and column when these includes are not found, but normally when committed the full path might not be correct.  The parser will never be able to find this full path in the current repo, given that is in fact, not there. 

To avoid confusion, we have added an option that allows the parser to ignore includes with full paths. In hopes that the parser can still be used, avoiding the includes to be processed. 

We expect over time, and with the use of the linter (that this library powers up) to trust relative paths within Fluent Bit configurations.